BC-34 # extract content-only functions
