+++ lib/create.js
@@ -12,6 +12,7 @@ const pify = require('pify');
 
 // local modules
 
+const content = require('./content.js');
 const deploy = require('./deploy');
 const resource = require('./resource');
 const promise = require('./utils/promise');
@@ -41,7 +42,7 @@ function newInteraction (options) {
     // follow mostly the same process as in pull
     .then(() => resource.writeInteraction({
       cwd,
-      data: resource.newInteraction({ name, type }),
+      data: content.newInteraction({ name, type }),
       name
     }))
     .then(() => {
